+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 *   (improvement) Add `_mojave-dragged-item` class on dragged item in sortable.
 *   (improvement) Add option to disable sortable hook easily (default is still "enabled").
+*   (bug) Warn when mounting a sortable directly on a table.
+
 
 5.9.3
 =====